+++ src/pages/index.tsx
@@ -1,4 +1,5 @@
-export { default as MainPage } from './main/page/MainPage';
-export { default as MyPage } from './myPage/page/MyPage';
-export { default as SignUpPage } from './signUp/page/SignUpPage';
-export { default as LoginPage } from './loginPage/page/LoginPage';
+export { default as MainPage } from "./main/page/MainPage";
+export { default as MyPage } from "./myPage/page/MyPage";
+export { default as SignUpPage } from "./signUp/page/SignUpPage";
+export { default as LoginPage } from "./loginPage/page/LoginPage";
+export { default as RankingPage } from "./ranking/page/RankingPage";
