+++ src/shared/index.tsx
@@ -1,5 +1,8 @@
-export { default as logoPath } from './ui/logo/logo.png';
-export { default as userPath } from './ui/user/user.png';
-export { pathItemMap } from './pathItems/pathItems';
+export { default as logoPath } from "./ui/logo/logo.png";
+export { default as userPath } from "./ui/user/user.png";
+export { pathItemMap } from "./pathItems/pathItems";
 // export { default as bootStyles } from './ui/styles/styles.scss';
-export { default as authStyle } from './ui/auth/Auth.module.css';
+export { default as authStyle } from "./ui/auth/Auth.module.css";
+export { default as rank1Path } from "./ui/ranking/medal_1.png";
+export { default as rank2Path } from "./ui/ranking/medal_2.png";
+export { default as rank3Path } from "./ui/ranking/medal_3.png";
